refactor: 디렉토리 구조 변경
--- 본문 ---
디렉토리 구조를 변경하였다.

--- 꼬리말 ---
refactor
