+++ src/main/java/com/sprint/project/findex/mapper/DashboardMapper.java
@@ -2,7 +2,8 @@ package com.sprint.project.findex.mapper;
 
 import com.sprint.project.findex.dto.dashboard.DashboardQueryDto;
 import com.sprint.project.findex.dto.dashboard.IndexPerformanceDto;
-import com.sprint.project.findex.repository.projection.DashboardRankingProjection;
+import com.sprint.project.findex.repository.dashboard.projection.DashboardRankingProjection;
+import java.text.DecimalFormat;
 
 public final class DashboardMapper {
 
@@ -16,12 +17,14 @@ public final class DashboardMapper {
     double versus = currentPrice - beforePrice;
     double fluctuationRate = versus * 100.0 / beforePrice;
 
+    double normalizedFR = Math.round(fluctuationRate * 100) / 100.0;
+
     return new IndexPerformanceDto(
         dto.indexInfoId(),
         dto.indexClassification(),
         dto.indexName(),
         versus,
-        fluctuationRate,
+        normalizedFR,
         currentPrice,
         beforePrice
     );
