+++ src/main/java/com/sprint/project/findex/service/AutoSyncConfigService.java
@@ -8,11 +8,10 @@ import com.sprint.project.findex.entity.AutoSyncConfig;
 import com.sprint.project.findex.global.exception.ApiException;
 import com.sprint.project.findex.global.exception.ErrorCode;
 import com.sprint.project.findex.mapper.AutoSyncConfigMapper;
-import com.sprint.project.findex.repository.IndexInfoRepository;
+import com.sprint.project.findex.repository.indexinfo.IndexInfoRepository;
 import com.sprint.project.findex.repository.autosyncconfig.AutoSyncConfigRepository;
 import com.sprint.project.findex.entity.IndexInfo;
 import java.util.List;
-import java.util.NoSuchElementException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
