+++ src/main/java/com/sprint/project/findex/service/DashboardService.java
@@ -10,9 +10,9 @@ import com.sprint.project.findex.entity.IndexInfo;
 import com.sprint.project.findex.global.exception.ApiException;
 import com.sprint.project.findex.global.exception.ErrorCode;
 import com.sprint.project.findex.mapper.DashboardMapper;
-import com.sprint.project.findex.repository.DashboardRepository;
-import com.sprint.project.findex.repository.IndexInfoRepository;
-import com.sprint.project.findex.repository.projection.DashboardRankingProjection;
+import com.sprint.project.findex.repository.dashboard.DashboardRepository;
+import com.sprint.project.findex.repository.indexinfo.IndexInfoRepository;
+import com.sprint.project.findex.repository.dashboard.projection.DashboardRankingProjection;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.sql.Date;
