+++ src/main/java/com/sprint/project/findex/service/IndexDataService.java
@@ -11,8 +11,8 @@ import com.sprint.project.findex.entity.SourceType;
 import com.sprint.project.findex.global.exception.ApiException;
 import com.sprint.project.findex.global.exception.ErrorCode;
 import com.sprint.project.findex.mapper.IndexDataMapper;
-import com.sprint.project.findex.repository.IndexDataRepository;
-import com.sprint.project.findex.repository.IndexInfoRepository;
+import com.sprint.project.findex.repository.indexdata.IndexDataRepository;
+import com.sprint.project.findex.repository.indexinfo.IndexInfoRepository;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Slice;
