+++ src/main/java/com/sprint/project/findex/service/IndexSyncService.java
@@ -14,8 +14,8 @@ import com.sprint.project.findex.global.entity.JobType;
 import com.sprint.project.findex.global.entity.ResultType;
 import com.sprint.project.findex.global.exception.ApiException;
 import com.sprint.project.findex.global.exception.ErrorCode;
-import com.sprint.project.findex.repository.IndexDataRepository;
-import com.sprint.project.findex.repository.SyncJobRepository;
+import com.sprint.project.findex.repository.indexdata.IndexDataRepository;
+import com.sprint.project.findex.repository.syncjob.SyncJobRepository;
 import java.net.URLEncoder;
 import java.nio.charset.StandardCharsets;
 import java.time.Duration;
