+++ src/main/java/com/sprint/project/findex/service/SyncJobService.java
@@ -10,8 +10,8 @@ import com.sprint.project.findex.dto.syncjob.SyncJobRequestQuery;
 import com.sprint.project.findex.entity.IndexInfo;
 import com.sprint.project.findex.entity.SyncJob;
 import com.sprint.project.findex.mapper.SyncJobMapper;
-import com.sprint.project.findex.repository.IndexInfoRepository;
-import com.sprint.project.findex.repository.SyncJobRepository;
+import com.sprint.project.findex.repository.indexinfo.IndexInfoRepository;
+import com.sprint.project.findex.repository.syncjob.SyncJobRepository;
 import com.sprint.project.findex.service.openapi.internal.PersistentWorker;
 import jakarta.servlet.http.HttpServletRequest;
 import java.time.LocalDate;
