+++ src/main/java/com/sprint/project/findex/service/openapi/internal/PersistentWorker.java
@@ -10,9 +10,9 @@ import com.sprint.project.findex.global.entity.JobType;
 import com.sprint.project.findex.global.entity.ResultType;
 import com.sprint.project.findex.mapper.IndexDataMapper;
 import com.sprint.project.findex.mapper.IndexInfoMapper;
-import com.sprint.project.findex.repository.IndexDataRepository;
-import com.sprint.project.findex.repository.IndexInfoRepository;
-import com.sprint.project.findex.repository.SyncJobRepository;
+import com.sprint.project.findex.repository.indexdata.IndexDataRepository;
+import com.sprint.project.findex.repository.indexinfo.IndexInfoRepository;
+import com.sprint.project.findex.repository.syncjob.SyncJobRepository;
 import com.sprint.project.findex.service.AutoSyncConfigService;
 import java.math.BigDecimal;
 import java.time.LocalDate;
